requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mlflow==2.8.0
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.25.0
psutil==5.9.6
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
zipp==3.17.0
